Updated install scripts

+++ scripts/uninstall.sh
@@ -1,15 +1,10 @@
 VIM=~/.vim_backup
 VIMRC=~/.vimrc_backup
 
-rm -rf ~/.vim ~/.vimrc ~/.config/coc
+rm -r ~/.vim ~/.vimrc ~/.config/coc
 
-if [ -d "$VIM" ]; then
-    mv ~/.vim_backup ~/.vim
-    echo ".vim_backup found, moving to .vim"
-fi
+[ -d "$VIM" ] && mv ~/.vim_backup ~/.vim  && echo ".vim_backup found, moving to .vim"
 
-if [ -f "$VIMRC" ]; then
-    mv ~/.vimrc_backup ~/.vimrc
-    echo ".vimrc_backup found, moving to .vimrc"
-fi
+[ -f "$VIMRC" ] && mv ~/.vimrc_backup ~/.vimrc && echo ".vimrc_backup found, moving to .vimrc"
+
